Get newest changes from Develop to feature/roles

+++ src/main/java/com/example/demo/domain/group/GroupServiceImpl.java
@@ -10,11 +10,26 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+/**
+ * GroupService - Service Layer for Group Entity
+ *
+ * <ul>
+ *     <li>To {@link GroupServiceImpl#findAll()} groups </li>
+ *     <li>To get group by ID {@link GroupServiceImpl#findById(UUID)}</li>
+ *     <li>To {@link GroupServiceImpl#saveGroup(Group)}</li>
+ *     <li>To {@link GroupServiceImpl#updateGroup(UUID, Group)}</li>
+ *     <li>To {@link GroupServiceImpl#deleteGroup(UUID)}</li>
+ * </ul>
+ *
+ * @author Remo Aeberli
+ */
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class GroupServiceImpl implements GroupService {
     private final GroupRepository groupRepository;
+    static final String GROUP_MATCHING = "Group matching {id:\"";
+    static final String NOT_EXISTING = "\"} does not exist";
 
     @Override
     public List<Group> findAll() {return groupRepository.findAll();}
@@ -24,7 +39,7 @@ public class GroupServiceImpl implements GroupService {
         if(groupRepository.existsById(id)) {
             return groupRepository.findById(id);
         } else {
-            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+            throw new InstanceNotFoundException(GROUP_MATCHING + id + NOT_EXISTING);
         }
     }
 
@@ -45,16 +60,16 @@ public class GroupServiceImpl implements GroupService {
             newGroup.setId(id);
             return groupRepository.save(newGroup);
         } else {
-            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+            throw new InstanceNotFoundException(GROUP_MATCHING + id + NOT_EXISTING);
         }
     }
 
     @Override
     public void deleteGroup(UUID id) throws InstanceNotFoundException {
-        if(groupRepository.existsById(id)) {
+        if(groupRepository.existsById(id) && groupRepository.findById(id).isPresent()) {
             groupRepository.deleteById(id);
         } else {
-            throw new InstanceNotFoundException("Group matching {id:\"" + id + "\"} does not exist");
+            throw new InstanceNotFoundException(GROUP_MATCHING + id + NOT_EXISTING);
         }
     }
 }
